deployement/requirements.txt
fastapi
uvicorn
google-cloud-storage
google-cloud-bigquery
tensorflow
pydantic